rename module so that pytype passes
