+++ rasa/cli/data.py
@@ -91,7 +91,7 @@ def split_nlu_data(args):
 
 def validate_files(args):
     from rasa.core.validator import Validator
-    from rasa.importers.basic import RasaFileImporter
+    from rasa.importers.rasa import RasaFileImporter
 
     loop = asyncio.get_event_loop()
     file_importer = RasaFileImporter(
